requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu